implements Configuration::validate
This method will return a list of zero or more ValidationError
instances. Currently there is two kinds of validation being performed:

 * bundle is setting a property that is not defined in master.
 * bundle is setting a property with a different type that was defined
   in master.

This will enforce users of YCB to make sure that master defines the
configuration schema, and all overwriting rules have to conform to the
types and not add unanticipated configurations.